+++ src/main/java/cn/edu/hdu/blog/repository/GuestRepository.java
@@ -3,5 +3,5 @@ package cn.edu.hdu.blog.repository;
 import cn.edu.hdu.blog.entity.dto.Guest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GuestRepository extends JpaRepository<Guest,Long> {
+public interface GuestRepository extends JpaRepository<Guest,Integer> {
 }
