+++ src/main/java/cn/edu/hdu/blog/repository/SettingRepository.java
@@ -3,5 +3,5 @@ package cn.edu.hdu.blog.repository;
 import cn.edu.hdu.blog.entity.dto.Setting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SettingRepository extends JpaRepository<Setting,Long> {
+public interface SettingRepository extends JpaRepository<Setting,Integer> {
 }
